+++ src/oop_ext/foundation/callback/_tests/test_priority_callback.py
@@ -1,8 +1,8 @@
-from oop_ext.foundation.callback import PriorityCallback
+from oop_ext.foundation.callback import PriorityCallback0
 
 
 def testPriorityCallback() -> None:
-    priority_callback = PriorityCallback()
+    priority_callback = PriorityCallback0()
 
     called = []
 
@@ -25,7 +25,12 @@ def testPriorityCallback() -> None:
     priority_callback.Register(OnCall2, priority=2)
     priority_callback.Register(OnCall3, priority=1)
     priority_callback.Register(OnCall4, priority=3)
-    priority_callback.Register(OnCall5, priority=2)
+    unregister5 = priority_callback.Register(OnCall5, priority=2)
 
     priority_callback()
     assert called == [3, 1, 2, 5, 4]
+
+    called.clear()
+    unregister5.Unregister()
+    priority_callback()
+    assert called == [3, 1, 2, 4]