+++ CHANGELOG.rst
@@ -1,7 +1,11 @@
 2.2.0 (UNRELEASED)
 ------------------
 
+* ``PriorityCallback`` now has type checking support, similar to ``Callback``, with type checked variants: ``PriorityCallback0``, ``PriorityCallback1``, etc (`#128`_).
+* ``UnregisterContext`` is now public, meant to be used in type annotations.
 * Added support for Python 3.10.
+
+.. _#128: https://github.com/ESSS/oop-ext/pull/128
 
 2.1.0 (2021-03-19)
 ------------------